+++ app/src/main/java/com/example/quickconnect/ChatAdapter.java
@@ -1,140 +1,63 @@
 package com.example.quickconnect;
-import android.content.Context;
+
 import android.os.Build;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.RequiresApi;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.quickconnect.databinding.ChatItemLayoutBinding;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.squareup.picasso.Picasso;
 
-import java.time.LocalDateTime;
-import java.time.chrono.ChronoLocalDate;
-import java.time.chrono.ChronoLocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 
-public class ChatAdapter extends RecyclerView.Adapter {
+public class ChatAdapter extends RecyclerView.Adapter<ChatAdapter.ChatViewHolder> {
 
-    private static final int VIEWTYPE_SENT = 1;
-    private static final int VIEWTYPE_RECIEVE = 2;
-    private String userId;
+    private List<Chat> dataList;
+    private final String userId;
 
-    private final List<Message> messageList;
-
-    public ChatAdapter(List<Message> messageList) {
-        this.messageList = messageList;
+    public ChatAdapter(List<Chat> dataList) {
+        this.dataList = dataList;
         userId = FirebaseAuth.getInstance().getCurrentUser().getUid();
     }
 
-    @Override
-    public int getItemViewType(int position) {
-	        Message message = messageList.get(position);
-
-	        if (message.getSenderId() == userId) {
-	            return VIEWTYPE_SENT;
-	        } else {
-	            return VIEWTYPE_RECIEVE;
-	        }
-	    }
-
     @NonNull
     @Override
-    public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view;
-        if (viewType == VIEWTYPE_SENT)
-        {
-            view = LayoutInflater.from(parent.getContext()).inflate(R.layout.message_item_me, parent, false);
-            return new MeViewHolder(view);
-        }
-        else if (viewType == VIEWTYPE_RECIEVE)
-        {
-            view = LayoutInflater.from(parent.getContext()).inflate(R.layout.message_item_other, parent, false);
-            return new ThemViewHolder(view);
-        }
-        return null;
+    public ChatViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.chat_item_layout, parent, false);
+        return new ChatViewHolder(view);
     }
 
     @RequiresApi(api = Build.VERSION_CODES.O)
     @Override
-    public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position) {
-        DateTimeFormatter dateFormat = DateTimeFormatter.ofPattern("MMM dd");
-        DateTimeFormatter timeFormat = DateTimeFormatter.ofPattern("hh:ss");
+    public void onBindViewHolder(@NonNull ChatViewHolder holder, int position) {
+        Chat chat = dataList.get(position);
 
-        Message message = messageList.get(position);
-        LocalDateTime messageTimestamp = message.getTimestamp();
-        String formattedTime = messageTimestamp.format(timeFormat);
-        String formattedDate = messageTimestamp.format(dateFormat);
-
-        if (holder.getItemViewType() == VIEWTYPE_SENT) {
-            MeViewHolder newHolder = (MeViewHolder) holder;
-            newHolder.message.setText(message.getText());
-            newHolder.time.setText(formattedTime);
-
-            if (position - 1 >= 0 && messageTimestamp.toLocalDate().isEqual(messageList.get(position - 1).getTimestamp().toLocalDate())) {
-                newHolder.date.setText(formattedDate);
-            } else {
-                newHolder.date.setVisibility(View.GONE);
-            }
-
-            if (message.getImage() != null) {
-                newHolder.image.setVisibility(View.VISIBLE);
-                Picasso.get().load(message.getImage()).fit().centerCrop().into(newHolder.image);
-            }
+        if (userId == chat.getCustomerId()) {
+            holder.binding.chatItemTitle.setText(chat.getSupportName());
         } else {
-            ThemViewHolder newHolder = (ThemViewHolder) holder;
-            newHolder.message.setText(message.getText());
-            newHolder.time.setText(formattedTime);
-
-            if (position - 1 >= 0 && messageTimestamp.toLocalDate().isEqual(messageList.get(position - 1).getTimestamp().toLocalDate())) {
-                newHolder.date.setText(formattedDate);
-            } else {
-                newHolder.date.setVisibility(View.GONE);
-            }
-
-            if (message.getImage() != null) {
-                newHolder.image.setVisibility(View.VISIBLE);
-                Picasso.get().load(message.getImage()).fit().centerCrop().into(newHolder.image);
-            }
+            holder.binding.chatItemTitle.setText(chat.getCustomerName());
         }
-
+        holder.binding.chatItemLastmsg.setText(chat.getCategory());
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("ddd hh:mm");
+        holder.binding.chatItemTime.setText(chat.getMessages().get(chat.getMessages().size() - 1).getTimestamp().format(formatter));
     }
 
     @Override
     public int getItemCount() {
-        return messageList.size();
+        return dataList.size();
     }
 
-    public static class MeViewHolder extends RecyclerView.ViewHolder {
-        public TextView message, date, time;
-        public ImageView image;
+    public class ChatViewHolder extends RecyclerView.ViewHolder {
+        private final ChatItemLayoutBinding binding;
 
-        public MeViewHolder(View itemView) {
+        public ChatViewHolder(@NonNull View itemView) {
             super(itemView);
-            message = itemView.findViewById(R.id.message_me_text);
-            date = itemView.findViewById(R.id.message_me_date);
-            time = itemView.findViewById(R.id.message_me_time);
-            image = itemView.findViewById(R.id.message_me_image);
-        }
-    }
-
-    public static class ThemViewHolder extends RecyclerView.ViewHolder {
-        public TextView message,date, time;
-        public ImageView image;
-
-        public ThemViewHolder(View itemView) {
-            super(itemView);
-            message = itemView.findViewById(R.id.message_them_text);
-            date = itemView.findViewById(R.id.message_them_date);
-            time = itemView.findViewById(R.id.message_them_time);
-            image = itemView.findViewById(R.id.message_them_image);
+            binding = ChatItemLayoutBinding.bind(itemView);
         }
     }
 }
